requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.0
replicate==0.15.4
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0

# Optional: للإنتاج
# whitenoise==6.5.0
# psycopg2-binary==2.9.7

# للواتساب API (اختياري - حسب الخدمة المستخدمة)
# twilio==8.2.0
# pywhatkit==5.4
